Fixed include path

+++ src/move.cpp
@@ -1,4 +1,4 @@
-#include "include/move.h"
+#include "move.h"
 
 bool is_super_effective(Move_t move, Pokemon_t* opponent) {
     return true;
@@ -10,5 +10,5 @@ float get_effectiveness(Move_t move, Pokemon_t* opponent) {
 }
 
 uint16_t calculate_damage(Move_t move, Pokemon_t* pokemon, Pokemon_t* opponent) {
-    return (((2 * pokemon->level) / 5 + 2) * move.power * (pokemon->attack / opponent->defense) / 50 + 2) * get_effectiveness(move, opponent);
+    return (uint16_t) ((((2 * pokemon->level) / 5 + 2) * move.power * (pokemon->attack / opponent->defense) / 50 + 2) * get_effectiveness(move, opponent));
 }
